made home button into its own component
