+++ client/src/components/HomeButton/HomeButton.js
@@ -1,0 +1,26 @@
+import React, {useCallback} from 'react'
+import './HomeButton.scss'
+
+import {useHistory} from 'react-router-dom'
+import Button from '../Button/Button'
+
+const HomeButton = props => {
+    const history = useHistory()
+    const goHome = useCallback(() => {
+        const to = `/`
+        history.push(to)
+    }, [history])
+
+    return (
+        <Button
+            {...{
+                label: 'Home',
+                transparent : true,
+                extraClass: 'back-btn',
+                onClick: goHome
+            }}
+        />
+    )
+}
+
+export default HomeButton
