+++ client/src/controllers/FourZeroFour/FourZeroFour.js
@@ -1,7 +1,13 @@
 import React from 'react'
+import HomeButton from '../../components/HomeButton/HomeButton'
 
 const FourZeroFour = props => {
-    return <div> Page not found </div>
+    return (
+        <div>
+            <HomeButton />
+            <h1>Page not found</h1>
+        </div>
+    )
 }
 
 export default FourZeroFour
