+++ client/src/controllers/Profile/Profile.js
@@ -1,29 +1,18 @@
 import React, {useState, useCallback} from 'react'
 import {connect} from 'react-redux'
-import { useHistory } from "react-router-dom";
 import './Profile.scss'
 
-import Button from '../../components/Button/Button'
 import YourStuff from './YourStuff/YourStuff'
 import OthersStuff from './OthersStuff/OthersStuff'
-
+import HomeButton from '../../components/HomeButton/HomeButton';
 
 const Profile = props => {
     const [yourStuff, isYourStuff] = useState(true)
     const {user} = props
-    const history = useHistory()
-    const goHome = useCallback(() => history.push(`/`), [history])
 
     return (
         <div className='profile'>
-            <Button
-                {...{
-                    label: 'Home',
-                    transparent : true,
-                    extraClass: 'back-btn',
-                    onClick: goHome
-                }}
-            />
+            <HomeButton />
 
             <h1 className='pageTitle'>{user?user.username:'Not Logged In'}</h1>
